Pulls reused functions from Generator into helper functions. Still need to figure out group imports
